requirements.txt
aiofiles==25.1.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
click==8.3.3
fastapi==0.136.1
greenlet==3.5.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.15
psutil==7.2.2
pydantic==2.13.4
pydantic_core==2.46.4
python-crontab==3.3.0
python-dotenv==1.2.2
PyYAML==6.0.3
schedule==1.2.2
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
